Infrastructure + Model Updates
Redid the parser to reduce the number of independent variables. Also captured the NFL schedule and saved it as a JSON file. In the infrastructure, I added resources for the VPC, S3, and DynamoDB. Worked on building out the Terraform script for the VPC. Mostly done with the exception of security rules.