Redid Terraform components
Renamed and added some additional DynamoDB tables. Also added a module for EventBridge. Changed the configuration for the S3 (no longer going to have it serve as a backup). On top of that, added a linear regression model to the ML module to see if it performed better than the random forest. It did not.